Move FollowWaypointsBtNavigator to nav2_bt_waypoint_follower
