+++ nav2_bt_waypoint_follower/CMakeLists.txt
@@ -1,19 +1,15 @@
 cmake_minimum_required(VERSION 3.5)
-project(nav2_bt_navigator)
+project(nav2_bt_waypoint_follower)
 
 find_package(ament_cmake REQUIRED)
 find_package(nav2_common REQUIRED)
 find_package(rclcpp REQUIRED)
 find_package(rclcpp_action REQUIRED)
 find_package(rclcpp_lifecycle REQUIRED)
-find_package(std_msgs REQUIRED)
-find_package(geometry_msgs REQUIRED)
-find_package(nav2_behavior_tree REQUIRED)
 find_package(nav_msgs REQUIRED)
 find_package(nav2_msgs REQUIRED)
-find_package(behaviortree_cpp_v3 REQUIRED)
-find_package(std_srvs REQUIRED)
 find_package(nav2_util REQUIRED)
+find_package(nav2_bt_navigator REQUIRED)
 find_package(tf2_ros REQUIRED)
 
 nav2_package()
@@ -22,41 +18,34 @@ include_directories(
   include
 )
 
-set(bt_navigator_library_name bt_navigator)
-set(navigate_to_pose_executable_name bt_navigator)
+set(executable_name bt_waypoint_follower)
 
-add_executable(${navigate_to_pose_executable_name}
-  src/navigate_to_pose/main.cpp
+add_executable(${executable_name}
+  src/main.cpp
 )
 
-set(library_name ${bt_navigator_library_name}_core)
+set(library_name ${executable_name}_core)
 
 add_library(${library_name} SHARED
-  src/bt_navigator.cpp
-  src/navigate_to_pose/bt_navigator.cpp
-  src/ros_topic_logger.cpp
+  src/bt_waypoint_follower.cpp
 )
 
 set(dependencies
   rclcpp
   rclcpp_action
   rclcpp_lifecycle
-  std_msgs
-  geometry_msgs
-  nav2_behavior_tree
   nav_msgs
   nav2_msgs
-  behaviortree_cpp_v3
-  std_srvs
   nav2_util
+  nav2_bt_navigator
   tf2_ros
 )
 
-ament_target_dependencies(${navigate_to_pose_executable_name}
+ament_target_dependencies(${executable_name}
   ${dependencies}
 )
 
-target_link_libraries(${navigate_to_pose_executable_name} ${library_name})
+target_link_libraries(${executable_name} ${library_name})
 
 ament_target_dependencies(${library_name}
   ${dependencies}
@@ -68,7 +57,7 @@ install(TARGETS ${library_name}
   RUNTIME DESTINATION bin
 )
 
-install(TARGETS ${navigate_to_pose_executable_name}
+install(TARGETS ${executable_name}
   RUNTIME DESTINATION lib/${PROJECT_NAME}
 )
 
@@ -76,15 +65,11 @@ install(DIRECTORY include/
   DESTINATION include/
 )
 
-install(DIRECTORY behavior_trees DESTINATION share/${PROJECT_NAME})
-
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
   ament_lint_auto_find_test_dependencies()
 endif()
 
 ament_export_include_directories(include)
-ament_export_libraries(${library_name})
-ament_export_dependencies(${dependencies})
 
 ament_package()
